% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_global.R
\name{assign_to_global}
\alias{assign_to_global}
\title{Assign to global environment}
\source{
https://stackoverflow.com/questions/28180989/r-cmd-check-found-the-following-assignments-to-the-global-environment
}
\usage{
assign_to_global(name, object, pos)
}
\arguments{
\item{name}{Character. Name to assign the object}

\item{object}{Object to assign to the name}

\item{pos}{Integer 1L}
}
\value{
function. Assignment to global environment
}
\description{
Assign to global environment
}
