% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_r.R
\name{convert_r}
\alias{convert_r}
\title{Industrialized dataset formats conversion}
\usage{
convert_r(mask_filepath, output_path)
}
\arguments{
\item{mask_filepath}{Entire file path of the excel mask}

\item{output_path}{Folder path where the converted datasets will be placed}
}
\description{
Industrialized dataset formats conversion
}
