% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_and_files.R
\name{csloc}
\alias{csloc}
\alias{current_script_location()}
\title{A wrapper of current_script_location()}
\usage{
csloc(x)
}
\description{
A wrapper of current_script_location()
}
\examples{
csloc()

}
