% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_script_location.R
\name{current_script_location}
\alias{current_script_location}
\title{RStudio only - Get exact path of the currently opened script}
\usage{
current_script_location()
}
\value{
Character 1L. A character vector indicating exact path of currently opened script
}
\description{
This function only works in RStudio and with an opened source document.
There is exceptionnaly no example and exact test since these are running in background sessions in checks.
}
