% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_import.R
\name{ext_tract}
\alias{ext_tract}
\title{Isolate the last level of an absolute file path}
\usage{
ext_tract(path, split = "/")
}
\arguments{
\item{path}{A valid R path}

\item{split}{Path levels delimiter}
}
\value{
A character vector corresponding to the last level of the path
}
\description{
Isolate the last level of an absolute file path
}
