% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_and_files.R
\name{folder_structure_replicate}
\alias{folder_structure_replicate}
\title{Replicate the folder structure of a given directory}
\usage{
folder_structure_replicate(dir, to)
}
\arguments{
\item{dir}{Path of directory which structure will be replicated}

\item{to}{Path of an output directory in which replicated structured will be placed}
}
\description{
Replicate the folder structure of a given directory
}
