% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{import_files}
\alias{import_files}
\title{Import datasets with a file selector}
\usage{
import_files(files = rChoiceDialogs::rchoose.files(), split = "/")
}
\arguments{
\item{files}{Where user interactive choices are stored}

\item{split}{Path separator, depending on user OS
The default value "/" is used on Linux
Use "\\\\" on Windows}
}
\value{
After completion, see the datasets imported in the Global Environment
}
\description{
Import datasets with a file selector
}
