% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{inspect}
\alias{inspect}
\title{Inspect a data frame}
\usage{
inspect(data_frame, nrow = FALSE)
}
\arguments{
\item{data_frame}{The data.frame to explore. Need to exist in the Global Environment.}

\item{nrow}{Logical. If TRUE, the number of observations of the dataset is rendered in addition.}
}
\value{
Variable list of the dataset and systematic informations for each variable.
}
\description{
Inspect a data frame
}
\examples{
inspect(diamonds)
}
