% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{inspect_vars}
\alias{inspect_vars}
\title{Inspect a collection of datasets}
\usage{
inspect_vars(input_path, output_path, output_label, considered_extensions)
}
\arguments{
\item{input_path}{Folder path of datasets to explore}

\item{output_path}{Folder path where the exploration output will be stored}

\item{output_label}{1-length character vector describing concisely the collection to explore}

\item{considered_extensions}{Character vector of extensions of datasets retained for exploration
in the input folder. Do not type the "." (dot) in it. E.g ("parquet" but not ".parquet")}
}
\value{
An excel file written on the computer containing exploration information
}
\description{
Inspect a collection of datasets
}
\examples{
saveRDS(cars, "./cars1.rds")
saveRDS(mtcars, "./cars2.rds")

# Code below illustrates how to use the function:
inspect_vars(input_path = ".", output_path = ".",
             output_label = "cardata", considered_extensions = "rds")

purrr::map(1:10, \(x)
           rio::import("./inspect_vars_cardata.xlsx",
                       sheet = x)) \%>\%
setNames(c("dims", "inspect_tot", "inspect_cars1", "inspect_cars2",
           "vars_detect", "vars_detect_everywhere", "vars_detect_not_everywhere",
           "vars_compclasses", "vars_compclasses_allsame", "vars_compclasses_not_allsame"))
           # code above illustrates all 10 sheets of the output

file.remove(c("./cars1.rds", "./cars2.rds", "./inspect_vars_cardata.xlsx"))

}
