% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_write.R
\name{inspect_write}
\alias{inspect_write}
\title{Inspect a data frame and write the output in an excel file}
\usage{
inspect_write(data_frame_name, output_path, output_label = NULL)
}
\arguments{
\item{data_frame_name}{Character 1L.
The data.frame to explore. Must exist in the Global Environment}

\item{output_path}{Character 1L. Folder path where to write the excel output}

\item{output_label}{Character 1L. Optional 1-length character vector to label output file.
If NULL, data_frame_name will also used as label}
}
\value{
See output xlsx file. Variable list of the dataset and systematic informations for each variable
}
\description{
Inspect a data frame and write the output in an excel file
}
\examples{
library(magrittr)
data(cars)
mydir <- file.path(tempdir(), "inspect_variants_tests_examples")
dir.create(mydir)
inspect_write(data_frame_name = "cars", 
              output_path = mydir, 
              output_label = "cars")
readxl::read_xlsx(file.path(mydir, "inspect_cars.xlsx"))
unlink(mydir, recursive = TRUE)
}
