% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ljoin_checks}
\alias{ljoin_checks}
\title{Perform a classical dplyr::left_join() and add check information related to join}
\usage{
ljoin_checks(ltable, rtable, ...)
}
\arguments{
\item{ltable}{Left data frame in the join}

\item{rtable}{Right data frame in the join}

\item{...}{Any other arguments of dplyr::left_join()}
}
\value{
Output of dplyr::left_join() with messages on number of observations
in left, right and joined data frames and list of common variables between
ltable and rtable
}
\description{
Perform a classical dplyr::left_join() and add check information related to join
}
