% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_r.R
\name{mask_convert_r}
\alias{mask_convert_r}
\title{Create an excel mask compatible with the convert_r() function}
\usage{
mask_convert_r(output_path, output_filename = "mask_convert_r.xlsx")
}
\arguments{
\item{output_path}{Folder path where the mask will be created}

\item{output_filename}{File name (with extension) of the mask}
}
\description{
Create an excel mask compatible with the convert_r() function
}
