% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_and_files.R
\name{path_level}
\alias{path_level}
\title{Extract a part of file path with levels}
\usage{
path_level(path, level)
}
\arguments{
\item{path}{Path to process}

\item{level}{Positive or negative integer. Level of the path, see example.}
}
\description{
Extract a part of file path with levels
}
\examples{
path_level("level_1/level_2/file.ext", 1)
path_level("level_1/level_2/file.ext", - 1)
path_level("level_1/level_2/file.ext", 0)

}
