% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_multiple.R
\name{replace_multiple}
\alias{replace_multiple}
\title{Replace character vector values using a correspondence approach}
\usage{
replace_multiple(input_vector, replacements, replace_all = FALSE)
}
\arguments{
\item{input_vector}{Character vector on which replacements take place}

\item{replacements}{Named character vector: replacement correspondences}

\item{replace_all}{Logical. If TRUE, stringr::str_replace_all() is used
instead of stringr::str_replace()}
}
\description{
Replace character vector values using a correspondence approach
}
\examples{
input <- c("one-one", "two-two-one", "three-three-two")

replace_multiple(input,
                 replacements =
                   c("one" = "1", "two" = "2",
                     "three" = "3"))

replace_multiple(input,
                 replacements =
                   c("one" = "1", "two" = "2",
                     "three" = "3"),
                 replace_all = TRUE)
}
