% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{serial_import}
\alias{serial_import}
\title{Import a collection of datasets into the Global Environment (serialized)}
\usage{
serial_import(file_paths, pos = globalenv())
}
\arguments{
\item{file_paths}{A character vector of valid absolute file paths of
datasets to import}
}
\value{
After completion, see the datasets imported in the Global Environment
}
\description{
Import a collection of datasets into the Global Environment (serialized)
}
