% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_prop}
\alias{table_prop}
\alias{base::table()}
\alias{base::prop.table()}
\title{Frequencies and proportions in one output}
\usage{
table_prop(..., margin = NULL, round = 3, noquote = FALSE)
}
\arguments{
\item{...}{Arguments passed to base::table()}

\item{margin}{The same argument as in base::prop.table()}

\item{round}{Number of digits after decimal in base::prop.table() output}

\item{noquote}{Logical. If TRUE, return an object of class noquote that provides better view of the output}
}
\value{
Frequencies with proportions in brackets, within a matrix
}
\description{
Combines base::table() and base::prop.table() outputs in a single one
}
\examples{
df <- data.frame(
                 "variable_1" = c("v1_1", "v1_1",
                                  "v1_2", "v1_2", "v1_2", "v1_2"),
                 "variable_2" = c("v2_1", "v2_1", "v2_1", "v2_1",
                                  "v2_2", "v2_2")
                                  )
table_prop(df$variable_1)
table_prop(df$variable_1, df$variable_2)
table_prop(df$variable_1, df$variable_2, margin = 2, noquote = TRUE)
}
