% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_compclasses_allsame.R
\name{vars_compclasses_allsame}
\alias{vars_compclasses_allsame}
\title{Variable class comparison - all of same type across all datasets}
\usage{
vars_compclasses_allsame(vars_compclasses_table)
}
\arguments{
\item{vars_compclasses_table}{Output of the vars_compclasses() function in this package.}
}
\value{
From vars_compclasses_table, extract type-consistent variables through all the datasets.
}
\description{
Variable class comparison - all of same type across all datasets
}
\examples{
library(magrittr)
vcompclasses_table <- vars_compclasses(c("cars", "mtcars"))
vars_compclasses_allsame(vcompclasses_table)
library(magrittr)
data(cars)
data(mtcars)
vcompclasses <- vars_compclasses(c("cars", "mtcars"))
vars_compclasses_allsame(vcompclasses)
}
