% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{vars_compclasses_not_allsame}
\alias{vars_compclasses_not_allsame}
\title{Variable class comparison - not all of same type across all datasets}
\usage{
vars_compclasses_not_allsame(vars_compclasses_table)
}
\arguments{
\item{vars_compclasses_table}{Output of the vars_compclasses() function in this package.}
}
\value{
From vars_compclasses_table, extract type-inconsistent variables through all the datasets.
}
\description{
Variable class comparison - not all of same type across all datasets
}
\examples{
vcompclasses_table <- vars_compclasses(c("cars", "mtcars"))
vars_compclasses_not_allsame(vcompclasses_table)

}
