% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{vars_detect}
\alias{vars_detect}
\title{Variable detection patterns}
\usage{
vars_detect(data_frames)
}
\arguments{
\item{data_frames}{The datasets to explore. Need to exist in the Global Environment}
}
\value{
Variable list and indicators of presences/absences across all inputted datasets.
}
\description{
Variable detection patterns
}
\examples{
vars_detect(c("cars", "mtcars"))
}
