% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{vars_detect_not_everywhere}
\alias{vars_detect_not_everywhere}
\title{Variable detection - inconsistent patterns}
\usage{
vars_detect_not_everywhere(vars_detect_table)
}
\arguments{
\item{vars_detect_table}{Output of the vars_detect() function in this package.
This object must exists into the Global Environment.}
}
\value{
From vars_detect_table, extract variables that are not always present through all the datasets.
}
\description{
Variable detection - inconsistent patterns
}
\examples{
vdetect_table <- vars_detect(c("cars", "mtcars"))
vars_detect_not_everywhere(vdetect_table)
}
