% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chars_manipulations.R
\name{chars_structure}
\alias{chars_structure}
\title{Get [:alpha:] / [:digit:] patterns from each symbol of character vector}
\usage{
chars_structure(input_vector)
}
\arguments{
\item{input_vector}{Character vector to process}
}
\value{
Character vector describing structure of each elements of input_vector,
see example
}
\description{
Get [:alpha:] / [:digit:] patterns from each symbol of character vector
}
\examples{
chars_structure(c("ABC123", "DE4F56", "789GHI"))

}
