% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_r.R
\name{convert_r}
\alias{convert_r}
\title{Industrialized dataset formats conversion}
\usage{
convert_r(mask_path, output_path)
}
\arguments{
\item{output_path}{Folder path where the converted datasets will be placed}

\item{mask_filepath}{Entire file path of the excel mask}
}
\description{
Industrialized dataset formats conversion
}
