% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_and_files.R
\name{current_script_location}
\alias{current_script_location}
\title{Get exact path of the currently opened script}
\usage{
current_script_location(x)
}
\value{
A character vector indicating exact path of currently opened script
}
\description{
Get exact path of the currently opened script
}
\examples{
current_script_location()

}
