% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupl_show.R
\name{dupl_sources}
\alias{dupl_sources}
\title{Illustrate sources of all duplicated values of a variable or a combination of variables}
\usage{
dupl_sources(data_frame, vars)
}
\arguments{
\item{data_frame}{Input data frame. Need to be in the Global Environment and has a data.frame class}

\item{vars}{A character vector of variable or combination of variables
from which duplicates are checked}
}
\value{
The part inputted data frame with all observations of duplicated values of
indicated variable or combination of variables
}
\description{
Illustrate sources of all duplicated values of a variable or a combination of variables
}
