% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_and_files.R
\name{file_extract}
\alias{file_extract}
\title{Extract filename given a filepath}
\usage{
file_extract(file_path, split = "/")
}
\arguments{
\item{file_path}{A 1L character vector. The file path from which file name will be extracted}

\item{split}{Indicates the path separator symbol}
}
\value{
Last level of filepath
}
\description{
Extract filename given a filepath
}
\examples{
library(magrittr)
file_extract(file_path = "level_1/level_2/file.ext",
                       split = "/")
}
