% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industtry-package.R
\docType{package}
\name{industtry-package}
\alias{industtry}
\alias{industtry-package}
\title{industtry: A Toolkit for Structured Datasets Exploitation}
\description{
This package proposes a set of functions that helps exploiting structured datasets (mostly data frames) with an industrialization approach. Industrialization here means applying as efficiently as possible the same procedure to any number of inputs as long as these inputs have an identified common structure.
}
\author{
\strong{Maintainer}: Daniel Rakotomalala \email{rakdanielh@gmail.com}

}
\keyword{internal}
