% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industtry-package.R
\docType{package}
\name{industtry-package}
\alias{industtry}
\alias{industtry-package}
\title{industtry: A toolkit for structured datasets exploitation}
\description{
A toolkit for structured datasets exploitation. A set of functions that helps exploiting structured datasets (data frames) with an industrialization approach. Industrialization here means applying as efficiently as possible the same procedure to any number of inputs as long as these inputs have an identified common structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielrak/industtry}
  \item Report bugs at \url{https://github.com/danielrak/industtry/issues}
}

}
\author{
\strong{Maintainer}: Daniel Rakotomalala \email{rakdanielh@gmail.com}

}
\keyword{internal}
