% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{inspect_vars}
\alias{inspect_vars}
\title{Inspect a collection of datasets}
\usage{
inspect_vars(input_path, output_path, output_label, considered_extensions)
}
\arguments{
\item{input_path}{Character 1L. Folder path of datasets to explore}

\item{output_path}{Character 1L. Folder path where the exploration output will be stored}

\item{output_label}{Character 1L. Describe concisely the collection to explore}

\item{considered_extensions}{Character. Extensions of datasets retained for exploration
in the input folder. Do not type the "." (dot) in it. E.g ("parquet" but not ".parquet")}
}
\value{
An excel file written on the computer containing exploration information
}
\description{
Inspect a collection of datasets
}
\examples{
mydir <- tempdir()
dir.create(mydir)

library(magrittr)
saveRDS(cars, file.path(mydir, "cars1.rds"))
saveRDS(mtcars, file.path(mydir, "cars2.rds"))

# Code below illustrates how to use the function:
inspect_vars(input_path = mydir, output_path = mydir,
              output_label = "cardata", 
             considered_extensions = "rds")

 purrr::map(1:10, \(x)
            rio::import(file.path(mydir, "inspect_vars_cardata.xlsx"),
                        sheet = x)) \%>\%
 magrittr::set_names(c("dims", "inspect_tot", "inspect_cars1", "inspect_cars2",
            "vars_detect", "vars_detect_everywhere", "vars_detect_not_everywhere",
            "vars_compclasses", "vars_compclasses_allsame", "vars_compclasses_not_allsame"))
            # code above illustrates all 10 sheets of the output

unlink(mydir, recursive = TRUE)
}
