% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_variants.R
\name{inspect_write}
\alias{inspect_write}
\title{Inspect a data frame and write the output in an excel file}
\usage{
inspect_write(data_frame_name, output_path, output_label = NULL)
}
\arguments{
\item{data_frame_name}{Character vector.
The data.frame to explore. Need to exist in the Global Environment}

\item{output_path}{Folder path where to write the excel output}

\item{output_label}{Optional 1-length character vector to label output file.
If NULL, data_frame_name will also used as label}
}
\value{
Variable list of the dataset and systematic informations for each variable
}
\description{
Inspect a data frame and write the output in an excel file
}
