% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importations.R
\name{parallel_import}
\alias{parallel_import}
\title{Import a collection of datasets into the Global Environment (parallelized)}
\usage{
parallel_import(file_paths)
}
\arguments{
\item{file_paths}{A character vector of valid absolute file paths of
datasets to import}
}
\value{
After jobs completion, see the datasets imported in the Global Environment
}
\description{
Import a collection of datasets into the Global Environment (parallelized)
}
