% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_move.R
\name{path_move}
\alias{path_move}
\title{Title}
\usage{
path_move(path_vector, path_separator = "/", move)
}
\arguments{
\item{path_vector}{Character vector. Vector of paths with equal number of levels}

\item{path_separator}{1L character vector. Path separator
(adapted to your OS for instance)}

\item{move}{1L integer.
If move > 0, outputs path till specified level in move.
If move < 0, remove the last specified level(s) in move.}
}
\value{
Character vector. Transformed vector of paths.
}
\description{
Title
}
\examples{
pvector <- c(
  "level_1/level_2/level_3/file_1.ext", 
  "level_1/level_2/level_3/file_2.ext"
)

path_move(path_vector = pvector, 
          path_separator = "/", 
          move = 1)

path_move(path_vector = pvector, 
          path_separator = "/", 
          move = 2)

path_move(path_vector = pvector, 
          path_separator = "/", 
          move = - 1)

path_move(path_vector = pvector, 
          path_separator = "/", 
          move = - 2)
}
