% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_import.R
\name{serial_import}
\alias{serial_import}
\title{Import a collection of datasets into the Global Environment (serialized)}
\usage{
serial_import(file_paths)
}
\arguments{
\item{file_paths}{Character. Vector of valid absolute file paths of
datasets to import}
}
\value{
After completion, see the datasets imported in the Global Environment
}
\description{
Import a collection of datasets into the Global Environment (serialized)
}
\examples{
library(magrittr)
mydir <- system.file("permadir_examples_and_tests/importations", package = "industtry")
lfiles <- list.files(mydir, full.names = TRUE) \%>\% 
  (\(x) x[stringr::str_detect(x, "\\\\.rds$")])

# Launch this code and check your global env, 
# cars.rds and mtcars.rds should be there: 
serial_import(lfiles)
}
