% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_prop}
\alias{table_prop}
\alias{base::table()}
\alias{base::prop.table()}
\title{Frequencies and proportions in one output}
\usage{
table_prop(..., margin = NULL, round = 3, noquote = FALSE)
}
\arguments{
\item{...}{Arguments passed to base::table()}

\item{margin}{The same argument as in base::prop.table()}

\item{round}{Number of digits after decimal in base::prop.table() output}

\item{noquote}{Logical. If TRUE, return an object of class noquote that provides better view of the output}
}
\value{
Frequencies with proportions in brackets
}
\description{
Combines base::table() and base::prop.table() outputs in a single one
}
