% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_vars.R
\name{vars_compclasses}
\alias{vars_compclasses}
\title{Collection-level variables types comparison}
\usage{
vars_compclasses(data_frames)
}
\arguments{
\item{data_frames}{The datasets to explore. Need to exist in the Global Environment}
}
\value{
Variable list and their respective types across all inputted datasets.
}
\description{
Collection-level variables types comparison
}
\examples{
vars_compclasses(c("cars", "mtcars"))
}
